whisperapp/requirements.txt
uvicorn
fastapi
numpy
python-multipart
openai-whisper
